+++ resources/lang/sk/dielo.php
@@ -29,6 +29,7 @@ return array(
     'item_attr_identifier'     => 'inventárne číslo',
     'item_attr_licence'        => 'licencia',
     'item_attr_place'          => 'geografická oblasť',
+    'item_attr_acquisition_date' => 'dátum akvizície',
     
     'more-items_related-artworks' => 'súvisiace diela',