+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Added
+- Remove query string from localized urls
 
 ## [1.5.0] - 2018-07-04
 ### Fixed