+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Fixed
+- Search endpoint of admin fixed
+
 ### Added
 - Remove query string from localized urls